\name{chemometrics}
\docType{data}
\alias{chemometrics}
\title{A chemometrics data set with FTIR spectrum data}
\description{
The \sQuote{chemometrics} data contains several analytes contents in a silage with the FTIR spectrum signal data. 
}
\format{The \sQuote{chemometrics} data set contains a response data frame \code{chemometrics} and a FTIR signal matrix \code{FTIR}:
 \describe{
 \item{analyte}{A data frame of content response.}
 \item{FTIR}{A matrix where each column corresponds to a wavenumber from the FTIR experiment.}
}
}
\keyword{datasets}