\name{plot.bigRR}
\alias{plot}
\alias{plot.bigRR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Print method for bigRR objects}
\description{
  A plot method for the object class bigRR
}
\usage{
\method{plot}{bigRR}(obj, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{obj}{A bigRR object }
  \item{alpha}{Transparency}
  \item{\dots}{ arguments not used }
}
\details{
  Plot bigRR objects.
}
