\name{print.bigRR}
\alias{print}
\alias{print.bigRR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Print method for bigRR objects}
\description{
A print method for the object class \code{bigRR}
}
\usage{
\method{print}{bigRR}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A bigRR object }
  \item{\dots}{ arguments not used }
}
\details{
  Provides a concise print of the elements of bigRR objects.
}
