\name{Arabidopsis}
\docType{data}
\alias{Arabidopsis}
\title{Arabidopsis thaliana data from Atwell et al. 2010 Nature}
\description{
The \sQuote{Arabidopsis} data contains a binary disease-related phenotype of Arabidopsis and an incidence matrix of genotypes from a 250K SNP chip. 
}
\format{ The \sQuote{Arabidopsis} data set contains a phenotype vector \code{y} and a genotype matrix \code{Z}:
 \describe{
 \item{y}{Binary vector of phenotype \emph{AvrRpm1} for 84 inbred lines.}
 \item{Z}{A 84 times 216130 matrix of genotypes of the 84 inbred lines, with two different homozygotes coded as -1 and 1.}
}
}
\source{Atwell, S., Y. S. Huang, B. J. Vilhjalmsson, G. Willems, M. Horton, et al., 2010 Genome-wide association study of 107 phenotypes in Arabidopsis thaliana inbred lines. \emph{Nature} 465: 627Ð631.}
\references{Atwell, S., Y. S. Huang, B. J. Vilhjalmsson, G. Willems, M. Horton, et al., 2010 Genome-wide association study of 107 phenotypes in Arabidopsis thaliana inbred lines. \emph{Nature} 465: 627Ð631.}
\keyword{datasets}