\name{bigRR-package}
\alias{bigRR-package}
\docType{package}
\title{
Fits big ridge regressions with special computational advantage for p>>n caeses.
}
\description{
The package fits ridge rigression. The shrinkage parameter (lambda) may be pre-specified or can be estimated. It gives possibilit to shrink any subset of parameters in the model. It has special computational advantage for the cases when the number of shrinkage parameter exceeds number of observations.
}
\details{
\tabular{ll}{
Package: \tab bigRR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-01-23\cr
License: \tab Unlimited\cr
LazyLoad: \tab yes\cr
}

}
\author{
Authors: Lars Ronnegard, Moudud Alam, Xia Shen\cr
Maintainer: Lars Ronnegard <lrn@du.se>
}
\references{
Ronnegard et al. (2010)
}

