\name{bigRR-package}
\alias{bigRR-package}
\docType{package}
\title{
Generalized Ridge Regression (with special advantage for p >> n cases)
}
\description{
The package fits (general) ridge regression for various distributions of response. The shrinkage parameters (lambdas) can be pre-specified or estimated using an internal update routine (fitting a heteroscedastic effects model). It gives possibility to shrink any subset of parameters in the model. It has special computational advantage for the cases when the number of shrinkage parameters exceeds the number of observations.
}
\details{
\tabular{ll}{
Package: \tab bigRR\cr
Type: \tab Package\cr
Version: \tab 1.3-2\cr
Date: \tab 2012-09-19\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{
Authors: Xia Shen, Moudud Alam, Lars Ronnegard\cr
Maintainer: Xia Shen <xia.shen@slu.se>
}

