\name{Chemometrics}
\docType{data}
\alias{Chemometrics}
\title{An Ethanol data set with FTIR spectrum data}
\description{
The \sQuote{Ethanol} data contains a response of ethanol content in a solution with the FTIR spectrum signal data. 
}
\format{The \sQuote{Ethanol} data set contains a response vector \code{ethanol} and a FTIR signal matrix \code{Z.FTIR}:
 \describe{
 \item{ethanol}{A vector of Ethanol content response.}
 \item{Z.FTIR}{A matrix where each column corresponds to a wavenumber from the FTIR experiment.}
}
}
\keyword{datasets}