\name{plot.bigRR}
\alias{plot}
\alias{plot.bigRR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Plot method for bigRR objects}
\description{
  A plot method for the objects of class "bigRR"
}
\usage{
\method{plot}{bigRR}(x, alpha = .36, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A "bigRR" object.}
  \item{alpha}{Transparency, ranged from 0 (completely transparent) to 1 (no transparency).}
  \item{\dots}{ arguments not used.}
}
\details{
  Plot the bigRR objects returned by \code{bigRR} or \code{bigRR.update}.
}
